history table recreated to avoid deletion conflicts on pets

+++ server/public/scripts/controllers/history.controller.js
@@ -1,31 +1,35 @@
-app.controller('HistoryController', ['$http', function($http) {
+app.controller('HistoryController', ['$http', function ($http) {
     let vm = this;
     vm.visits = [];
 
-    vm.getHistory = function() {
+    vm.getHistory = function () {
         $http({
             method: 'GET',
             url: '/history'
-        }).then(function(response) {
+        }).then(function (response) {
             vm.visits = response.data;
             vm.refineTimes(vm.visits);
-        }).catch(function(error){
+        }).catch(function (error) {
             console.log('error:', error);
             alert('failed to connect with server');
         });
     }//end getHistory
 
-    vm.deleteItem = function(item) {
-        $http({
-            method: 'DELETE',
-            url: '/history',
-            params: {id: item.id}
-        }).then( function (response){
-            vm.getHistory();
-        }).catch( function(error) {
-            console.log('error:', error);
-            alert('failed to delete item');
-        });
+    vm.deleteItem = function (item) {
+        if (item.out === 'ongoing') {
+            alert('you must checkout the pet first');
+        } else {
+            $http({
+                method: 'DELETE',
+                url: '/history',
+                params: { id: item.id }
+            }).then(function (response) {
+                vm.getHistory();
+            }).catch(function (error) {
+                console.log('error:', error);
+                alert('failed to delete item');
+            });
+        }
     }//end deleteItem
 
     vm.refineTimes = function () {
@@ -42,6 +46,6 @@ app.controller('HistoryController', ['$http', function($http) {
             }
         }
     }//end refineTimes
-    
+
     vm.getHistory();
 }]);
